Update patch set 2
Patch Set 2:

In this case, it sounds like the flag being added to the build system will result in system images that don't work. I'd suggest that instead of this change, you talk to the owners of the APKs in question to modify their process and provide you with the APKs in a form which is suitable for installing via Package Manager. Alternatively, if the APKs come from the source tree as prebuilts, you can simply obtain it from the source tree.

Patch-set: 2
Reviewer: Alex Klyubin <1018773@85c56323-6fa9-3386-8a01-6480fb634889>