Update patch set 5
Patch Set 5:

> Why is this limited to hardware/qcom?

Projects in this location are the first clear instance I identified of device-specific projects that should not install to system.

> Why is it deferring the warning until the build runs?

Not sure I understand. How is it doing this?

> Why does this need to be different than the link type checks?

This are binaries that are being installed to the wrong location. If it's a self contained device-specific binary then you will not catch it with the link type checking.

Patch-set: 5
Reviewer: Diego Wilson <1124149@85c56323-6fa9-3386-8a01-6480fb634889>