Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Why is this limited to hardware/qcom?
> 
> Projects in this location are the first clear instance I identified of device-specific projects that should not install to system.

There are many other locations that very well may contain mixed generic and device-specific code (especially under vendor/...). Any sort of hardcoded isn't scalable, and any other indication that this is device specific should just force the module to be installed correctly (currently, all we've got is LOCAL_PROPRIETARY_MODULE, but that needs to change)

> > Why is it deferring the warning until the build runs?
> 
> Not sure I understand. How is it doing this?

We know the install location and the makefile location while we parse the makefiles -- this could be done as a $(warning) while reading the makefiles instead of defining a whole rule.

> > Why does this need to be different than the link type checks?
> 
> This are binaries that are being installed to the wrong location. If it's a self contained device-specific binary then you will not catch it with the link type checking.

Good point. We've had build/make/core/tasks/vendor_module_check.mk for the previous version of this check, which catches everything, but relies on the fact that vendor/(src tree) == /vendor(partition).

Patch-set: 5
Reviewer: Dan Willemsen <1074628@85c56323-6fa9-3386-8a01-6480fb634889>