Update patch set 5
Patch Set 5:

> There are many other locations that very well may contain mixed generic and device-specific code (especially under vendor/...). Any sort of hardcoded isn't scalable, and any other indication that this is device specific should just force the module to be installed correctly (currently, all we've got is LOCAL_PROPRIETARY_MODULE, but that needs to change)

You are right in that the process of identifying device-specific modules is mostly manual. My idea is to create a grey list of projects that need migration since the process will likely be a cross-team effort. I agree that it's not ideal and I'm open to any alternatives you may have in mind.

If you don't believe the effort deserves to be supported by the mainline build system then I'll just share the patch with each Android team involved individually.

Patch-set: 5
Reviewer: Diego Wilson <1124149@85c56323-6fa9-3386-8a01-6480fb634889>