Add aosp_arm64_ab for generic system image for A/B-ready devices
aosp_arm64_ab is added as a product to build a generic system image for
all devices with ARM64 and A/B ready partitions. For now, it only
supports sailfish/marlin devices because following items are not yet
split.

- fstab
- overlays
- sepolicy

Bug: 35653062
Test: build and flash it to sailfish (or marlin)

$ source build/envsetup.sh; choosecombo 1 sailfish userdebug
$ ENABLE_EARLY_MOUNT=true ENABLE_TREBLE=true m -j 80
$ fastboot flashall
(For now, the two ENABLE_* are required to install  *.rc and *.prop
files into vendor partition. This restriction will be removed in the
future.)

$ source build/envsetup.sh; choosecombo 1 aosp_arm64_ab userdebug
$ m -j 80 systemimage
$ fastboot flash system out/target/product/generic_arm64_ab/system.img
$ fastboot -w reboot

Change-Id: Ia91163f2d51e90a488c9451cba23242887ea82ba