Update patch set 1
Patch Set 1:

NO, do NOT run logcat in the logd domain!!!!

Special case, logcat is run in logpersist domain, on user builds landing on the emulator. The logpersist domain restricts how much the logcat is allowed to do, used as a jail to ensure that no data leaks off the platform that we do not expect.

A _lot_ of what logpersist can access is turned off on user builds, it is IMHO OK to add emulator-specific perms, carefully reviewed for security issues.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>