Update patch set 5
Patch Set 5:

Not covered here: we need/want a mechanism where if fs_config_dirs is empty, that fs_config_files automagically adds to the file list system/<partition>/etc/fs_config_dirs as mode 0000 (unreadable) to limit the time-impact of the empty file. Currently we do ask that the developer takes this initiative. The alternative is to not request the file to be produced if it is not needed.

Patch-set: 5
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0