Update patch set 2
Patch Set 2: Code-Review-1

After some research, this is the algorithm I chose to use for automatic creation:

Sensitive to the following board definitions:
    
    BOARD_VENDORIMAGE_FILE_SYSTEM_TYPE
    BOARD_OEMIMAGE_FILE_SYSTEM_TYPE
    BOARD_ODMIMAGE_FILE_SYSTEM_TYPE
    
Which when defined and set means the images are generated for vendor, oem or odm respectively.
    
Sensitive to the following definitions:
    
    TARGET_OUT
    TARGET_OUT_VENDOR
    TARGET_OUT_OEM
    TARGET_OUT_ODM
    
When when not unique, suppresses vendor, oem or odm respectively in that order of precedence.

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Presubmit-Ready=0