Update patch set 11
Patch Set 11:

Is this all really necessary to enforce capabilities in the file system if we can override them with the init.rc files? We still need both methods to be able to control file mode and ownership (and capabilities no so much)?

Patch-set: 11
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0