Update patch set 6
Patch Set 6:

(3 comments)

Most of what this patch is about is consolidating the existing documentation into one spot. I guess I could adjust the commit message accordingly if you like. If we leave the canonical location for documentation as source.android.com, then we can't easily change the documentation quickly, validate/ensure that the documentation applies to the version it claims to apply to, or encourage the developer who is adding a feature to also document that feature while it's still fresh in their mind.

The list of individual commonly-used targets seem to me to be less useful than the other parts of README.md (or at least it depends on them), which explain the syntax of invoking a build, the types of targets, where targets come from, and what flags are.

It would seem reasonable to me to extract the common targets into a separate file called common_targets.md, and for `make help` to mention the existence of both `README.md` and of `common_targets.md` and to let the user decide whether they're more interested in basic concepts or in a list of targets. Is that acceptable?

For a proper longterm solution for discovering targets (assuming that the hardcoded list of targets becomes unwieldy), I imagine we'll either explain to the user how to discover targets, or even better mention a tool, once we've established a place to mention it (sort of like the .json file generator you mentioned).

Which targets are "framework-specific ones", and/or which ones are "the more useful generic ones"? (or does it not matter if we extract them all into common_targets.md)

Patch-set: 6
Reviewer: Jeff Gaston <1137668@85c56323-6fa9-3386-8a01-6480fb634889>