Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I suspect a CL from this topic broke compilation of the target "docs" on two different branches:
> 
> vendor/unbundled_google/libraries/openglhelper/jni/gl_code.cpp:3:10: fatal error: 'nativehelper/jni.h' file not found
> #include <nativehelper/jni.h>
> ^~~~~~~~~~~~~~~~~~~~
> 1 error generated.

Can you please fix or advise on the right fix? I'll refrain from rollback for now, until someone complains. Is the right fix to merely drop the line
  #include <nativehelper/jni.h>
since the commit message of this CL seems to suggest that the type is now available without the need for an #include?

Patch-set: 2