droiddoc: Fix trailing ':' in -classpath, -sourcepath values
full_java_lib_deps has trailing whitespace if $(LOCAL_CLASSPATH)
is empty. Because the calculation of PRIVATE_CLASSPATH was
missing a $(strip ...), this resulted in an (unintentional)
trailing ':' in the values of the -classpath and -sourcepath
arguments to the javadoc tool, which that tool is documented
to interpret as '.' (the working directory).

This CL converts the expression to a call to normalize-path-list,
from definitions.mk, which contains the $(strip ...):

define normalize-path-list
$(subst $(space),:,$(strip $(1)))
endef

After this CL, an empty $(LOCAL_CLASSPATH) no longer gets
misinterpreted as the current working directory.

This issue was minor (it made no difference in practice).

Test: Treehugger
Bug: 62049770

Change-Id: Ia0e3e5657d0fa057fe998515f34bc7b8df5f6f16