Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> I'm sure there's a more compact way of doing it, and perhaps clearer, but Colin can point that out if there is.

Simplified.

I had looked for ways to express "if (EXPERIMENTAL_USE_OPENJDK9 is set *and* ANDROID_JAVA_HOME is set)" so I don't have to duplicate, but couldn't find any.

Anyway, if EXPERIMENTAL_USE_OPENJDK9 is set but ANDROID_JAVA_HOME isn't (or doesn't point to a valid toolchain), then that is an error, so we don't really care about what happens. I've simplified the CL now by not accounting for that case. In patch set #1, the behavior would have been to use the bundled OpenJDK 8; in patch set #2, the build will likely fail, because there probably won't be a /bin/{javac,jar,javadoc}, which seems like the better behavior.

Patch-set: 1