Update patch set 3
Patch Set 3:

Makes sense given the lack of presubmit coverage.  It would be nice if we could verify the list for non-coverage builds and then fail early to avoid accumulating references to classes that don't exist, but probably not worth the effort.

Patch-set: 3
Reviewer: Colin Cross <1002751@85c56323-6fa9-3386-8a01-6480fb634889>