Copy resources from classes.jar when using AAPT2
When building with javac but without AAPT2 the final package is
created by copying classes.jar, deleting the class files, and
adding in the dex files.  This leaves any non-class java resources
in the final package.  When using AAPT2, the final package is
created from AAPT2's package-res.apk, so the java resources have
to be added back into the package manually.

Bug: 64272170
Test: m -j SystemUITests, check that the apk contains mockito-extensions/*
Change-Id: I6649f7f2fbf63ff5001f4731e22960f857726c70