Update patch set 7
Patch Set 7:

> Did this perhaps break compilation for some armv7 variants?
 > 
 > The dashboard is green, but I get:
 > $ make cts
 > build/core/combo/TARGET_linux-arm.mk:43: kryo is armv8-a.
 > build/core/combo/TARGET_linux-arm.mk:45: TARGET_2ND_ARCH_VARIANT,
 > armv7-a-neon, ignored! Use armv8-a instead.
 > [...]
 > cts/tests/tests/os/jni/android_os_cts_CpuInstructions.cpp:88:20:
 > error: instruction requires: armv7 or earlier
 > asm volatile ( "swp r0, r0, [%0]" : "+r"(ptr) : : "r0" );
 > ^
 > <inline asm>:1:2: note: instantiated into assembly here
 > swp r0, r0, [r1]
 > ^

I think so. It's because Kyo is not an armv7-a variant. It implements armv8-a ISA and the old swp instruction is obsolete (removed) in armv8-a!
Where is Kyro specified (as the TARGET_2ND_CPU_VARIANT)? You might want to change it to something like cortex-a15 or (more general, but less performant) "generic".

Patch-set: 7
Reviewer: Isaac Chen <1139803@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0