Update patch set 1
Patch Set 1:

It's not clear to me whether any better solution than this revert was found during the MTV business day, but since there is no evidence for that that I can see, and the build is still broken for me, I plan to go ahead and merge this revert.

I've manually verified that patching this CL topic fixes the build breakage that I observe.

Patch-set: 1
Reviewer: Tobias Thierer <1103203@85c56323-6fa9-3386-8a01-6480fb634889>