Fix unbundled jack builds
Recent change Icbc8987468f237aa351708b33c3abe5ea3d1816e was
unconditionally adding core-oj and core-libart as the
bootclasspath for all jack compiles, which is incorrect and
was breaking unbundled builds that don't have sources for
core-oj and core-libart.  Only add them when LOCAL_SDK_VERSION
is not set.

Test: unbundled build
Change-Id: I6bab7e36e8178384f13cf9df8183a53fcb4b23f8