Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm not trying to add work to William's plate, but it seems like we could solve this more generally by nuking every environment variable by default, and explicitly setting all the environment variables we expect. Any use of a user supplied environment variable by the build system could make the build non-reproducible. 

Yes, I definitely want that to happen soon, it's a large hole in our reproducibility/reliability story. Unfortunately we don't really know which environment variables people are using right now, and can't really make accesses errors, so it's not going to be a quick fix.

We can make things better in makefiles, since most of the environment variable access already go through kati, which tracks all of them. But at the same time, undefined variable accesses try to read the environment too, so there are a lot more accesses.

To some extent, LD_PRELOADing a getenv/etc wrapper may be able to get an approximate answer too.

> I'm also fine with just addressing this particular environment variable too.

Yeah, in this case I'd just handle this variable, especially since I'd like to see if we could clear/set the LANG/LC_* variables, except keep LC_MESSAGES so that error messages are still translated as much as possible. (At least the documentation implies that that should work, but I don't know if the implementations actually respect that)

We'll be working on the larger environment problem soon.

Patch-set: 1
Reviewer: Dan Willemsen <1074628@85c56323-6fa9-3386-8a01-6480fb634889>