Update patch set 5
Patch Set 5:

Dan, Colin, friendly ping on this?

Note that once b/30972906 stabilizes, we expect to move away from that new product variable and be explicit in the Makefile. That will require more work, which is the reason we're going with this simple workaround now.

Patch-set: 5
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>