Update patch set 2
Patch Set 2:

The input to jacoco is classes-desugar.jar . jacoco was failing on a jar within classes-desugar.jar, whereas I suspect desugar wasn't recursing into the subjar or was ignoring the corruption.

I'm not sure what Jack was doing differently. It wouldn't surprise me if it wasn't recursing into the subjar either.

I'd assumed that tests/resources was a normal directory to have exist in classes-desugar.jar , although on closer inspection that turns out to be incorrect - I only found it in frameworks/base/core/tests .

Also, I discovered that the one case where this was happening went away when https://android-review.googlesource.com/c/platform/libcore/+/516460 was merged.

So anyway, if it sounds good to everyone here, my inclination is to just abandon this patch now that it's no longer needed

Patch-set: 2
Reviewer: Jeff Gaston <1137668@85c56323-6fa9-3386-8a01-6480fb634889>