Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Actually IMAGES/{vbmeta,recovery}.img are correcting in the resulting target-files.zip file ... so I guess 'm dist' works correctly.

That's added very recently in https://android-review.googlesource.com/c/platform/build/+/537309. I overlooked the problem with 'make' by then, so it only addressed the issue with 'make dist'.

But there's a mismatching logic between the Makefile in this CL and the releasetools script. The releasetools script uses the same set of parameters to build boot.img and recovery.img, while in this CL it uses a separate set of flags (BOARD_AVB_RECOVERY_*). We should resolve the inconsistency.

Patch-set: 1
Reviewer: Tao Bao <1056365@85c56323-6fa9-3386-8a01-6480fb634889>