Update patch set 1
Patch Set 1:

The logic seems wrong to me; in my understanding, this CL tries to avoid a "free" command in the middle of two "stash" commands with the same SHA.
Example:
1. stash 0x100 into A;
2. use A for some operations;
3. free A;      ---> (avoid this cmd)
...
4. stash 0x7100 into A;
5. use A for some operations;
6. free A;

But in the current implementation, we skip stashing the blocks if a stash with the same SHA exists, even if the block contents are different (i.e. a hash collision). As a result, we'll use the stash 0x100 during the update while we actually expect 0x7100.
And we should keep the current behavior about stash overwriting, since the source blocks to stash may no longer be intact for interrupted updates.

Patch-set: 1
Reviewer: Tianjie Xu <1088085@85c56323-6fa9-3386-8a01-6480fb634889>