Update patch set 3
Patch Set 3:

The existing behavior for mmm (failing) seems right to me and the new behavior (walking up to an Android.mk) for mmm seems wrong to me because the latter silently changes the meaning of what the user asked for (whereas the former should just cause the user to change their command-line argument (do we want to display an error message listing the closest Android.mk?)). However, I only decided that after implementing it, and I'm switching teams tonight anyway, so here is what I implemented, and everyone else can decide what they want to do about it.

I didn't get a chance to exhaustively test it but based on a little bit of testing it seems to me that it implements what was requested, at least.

Patch-set: 3
Reviewer: Jeff Gaston <1137668@85c56323-6fa9-3386-8a01-6480fb634889>