Update patch set 1
Patch Set 1: Presubmit-Verified-1

=== Finished presubmit run: L52800000128114003 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/569163?revision=1




TreeHugger finished with: 18 passed, 2 failed.

WARNING: The tip-of-tree(4503807) for aosp-master/aosp_arm64-eng may be broken.
The presubmit build passes with the last-known-good-build(4503484), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



FAIL: Failed to build aosp-master/aosp_x86-eng/4503807 (ToT). See https://android-build.googleplex.com/builds/pending/P5098152/aosp_x86-eng/latest. 
Note: This is a checkbuild (go/checkbuild). It will take longer and could uncover errors not found in other builds.



FAIL: Failed to build git_master/errorprone_fastbuild3e_linux/4503813 (ToT). See https://android-build.googleplex.com/builds/pending/P5098167/errorprone_fastbuild3e_linux/latest

WARNING: The tip-of-tree(4503804) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4503524), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4503804) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4503524), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4503813) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4503640), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.







WARNING: The tip-of-tree(4503767) for git_mirror-aosp-master-with-vendor/aosp_sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4503521), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 1
Tag: autogenerated:TreeHugger:Verify-Fail
Reviewer: Treehugger Robot <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0
Label: Presubmit-Verified=-1