Update patch set 3
Patch Set 3:

I am also somewhat concerned about what system resources this needs -- I saw from the internal logs that you're failing acquire display resources, which is correct. Really we prefer tests not to be in the build, but for well-behaved tests it's not always a problem. If you're relying on system resources it's going to be a problem though.

Patch-set: 3
Reviewer: Dan Willemsen <1074628@85c56323-6fa9-3386-8a01-6480fb634889>