Use proper make variable for vendor directory
Usually, we use $(TARGET_COPY_OUT_VENDOR) in make file to refer to
the vendor directory on the device.

In target/product/emulator.mk on AOSP, this make variable is used,
but in internal master, "vendor" is used! This would cause some
automerger conflict when new CLs are commited to AOSP.

Test: $ lunch aosp_x86-userdebug; m -j; emulator
Change-Id: I18667ac76d0ba3000aa361051d02e4cbb62a4f2f