Update patch set 2
Patch Set 2:

We also call this function immediately before/after calls to uncompress-dex -- should the conditionals be fixed there so that we don't even call this function if we're uncompressing the dex files? Or was the problem here being hit in a different way?

Patch-set: 2
Reviewer: Dan Willemsen <1074628@85c56323-6fa9-3386-8a01-6480fb634889>