Update patch set 14
Patch Set 14:

Jaekyun, we (Android Things team) are actually following this CL because we're hoping to use this, so if we need to do some work to cleanly move from the legacy approach to this, we can make it happen.

Adding Sen and Steve who will likely handle that.

Patch-set: 14
CC: Sen Jiang <1079733@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Steve Fung <1076778@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Bryan Henry <1178083@85c56323-6fa9-3386-8a01-6480fb634889>