Update patch set 1
Patch Set 1:

Is there an example for how to forbid LOCAL_PATCH_MODULE for targets that are not tests? It might be needed for non-tests in future, but currently the only use case that I know about are test targets (and specifically, only two examples). Any suggestions on whether/how we should limit use of this feature to stop people figuratively shooting themselves in the foot?

Patch-set: 1
Reviewer: Tobias Thierer <1103203@85c56323-6fa9-3386-8a01-6480fb634889>