Update patch set 4
Patch Set 4: Code-Review-2

As discussed in the Fall, changes to accommodate the specific requirement for this project are to be done in a downstream branch, not affecting master (mainline Android).  If wifi is not supported on a device, the overlay should be configured so the wifi service is not started.

Patch-set: 4
Reviewer: Rebecca Silberstein <1081395@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2