Add ro.vendor.product.cpu.* to vendor/build.prop
CPU/ABIs related information is already in system/build.prop. It
should also be included in the vendor image so it can still be
examined outside of a device, build environment, or $OUT directory,
by a standalone utility to check its basic characteristics and if
it's compatible with some specific system image (usually GSI).

Bug: 72079894
Test: Built and verified aosp_x86(_64):
      # On host
      $ grep ro.vendor.product.cpu $OUT/vendor/build.prop
      # On device
      $ getprop | grep ro.vendor.product
      $ su
      # grep ro.vendor.product /vendor/build.prop

Change-Id: Ic9ac8065d5983710840e33a805f982de75d86ce3